add nographic execution mod for debugging and copy/paste + add doc for MemoryMapEntry shape + Modify get_block_number to correct size (int16_t) + Change types length (because of -m32, long int is not 8 bytes) + Change loop for Memory Mapping (in bootloader.asm) + Change value of memory_map_count, equ instead of dw (asm subtility) + Remove unused interrupt 0x20

+++ include/types.hpp
@@ -2,12 +2,12 @@
 
 #define NULL 0
 
-typedef long int int64_t;
+typedef long long int int64_t;
 typedef int int32_t;
 typedef short int16_t;
 typedef char int8_t;
 
-typedef unsigned long int uint64_t;
+typedef unsigned long long int uint64_t;
 typedef unsigned int uint32_t;
 typedef unsigned short uint16_t;
 typedef unsigned char uint8_t;
